requirements.txt
streamlit==1.28.0
pypdf==3.14.0
langchain==0.1.0
faiss-cpu==1.7.4
google-generativeai==0.3.0
python-dotenv==1.0.0
numpy==1.24.4
